Split Sempal native runtime adapter modules
